move bottom nav to another component
