+++ frontend/src/App.tsx
@@ -1,47 +1,14 @@
-import { useState } from 'react';
+import { FC } from 'react';
 import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom';
 
-import ArchiveIcon from '@mui/icons-material/Archive';
-import FavoriteIcon from '@mui/icons-material/Favorite';
-import RestoreIcon from '@mui/icons-material/Restore';
-import {
-  AppBar,
-  Toolbar,
-  Typography,
-  Container,
-  Paper,
-  BottomNavigation,
-  BottomNavigationAction,
-  Alert,
-  Snackbar,
-  Slide,
-  Grid,
-} from '@mui/material';
+import { AppBar, Toolbar, Typography, Container, Grid } from '@mui/material';
 
 import EditPost from './components/EditPost';
 import PostForm from './components/PostForm';
 import PostList from './components/PostList';
+import BottomNav from './components/layout/BottomNavigation';
 
-function SlideTransition(props) {
-  return <Slide {...props} direction="down" />;
-}
-
-const App: React.FC = () => {
-  const [value, setValue] = useState<number>(0);
-  const [open, setOpen] = useState<boolean>(false);
-
-  const handleChange = (event, newValue) => {
-    setValue(newValue);
-    setOpen(true);
-  };
-
-  const handleClose = (event, reason) => {
-    if (reason === 'clickaway') {
-      return;
-    }
-    setOpen(false);
-  };
-
+const App: FC = () => {
   return (
     <Router>
       <AppBar sx={{ position: 'fixed', top: 0, left: 0, right: 0 }}>
@@ -58,11 +25,7 @@ const App: React.FC = () => {
           </Typography>
         </Toolbar>
       </AppBar>
-      <Grid
-        role="main"
-        sx={{ padding: '80px 1rem', minHeight: '100vh', bgcolor: '#0d1117', color: '#e6edf3' }}
-        // lg={{ maxWidth: 'unset' }}
-      >
+      <Grid role="main" sx={{ padding: '80px 1rem', minHeight: '100vh', bgcolor: '#0d1117', color: '#e6edf3' }}>
         <Container>
           <Routes>
             <Route path="/" element={<PostList />} />
@@ -70,23 +33,7 @@ const App: React.FC = () => {
             <Route path="/edit/:id" element={<EditPost />} />
           </Routes>
         </Container>
-        <Paper sx={{ position: 'fixed', bottom: 0, left: 0, right: 0 }} elevation={3}>
-          <BottomNavigation showLabels value={value} onChange={handleChange} sx={{ background: '#121212' }}>
-            <BottomNavigationAction sx={{ color: '#e6edf3' }} label="Aktualności" icon={<RestoreIcon />} />
-            <BottomNavigationAction sx={{ color: '#e6edf3' }} label="Ulubione" icon={<FavoriteIcon />} />
-            <BottomNavigationAction sx={{ color: '#e6edf3' }} label="Archiwum" icon={<ArchiveIcon />} />
-          </BottomNavigation>
-        </Paper>
-        <Snackbar
-          open={open}
-          autoHideDuration={2500}
-          onClose={handleClose}
-          TransitionComponent={SlideTransition}
-          anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
-          style={{ marginTop: '4rem' }}
-        >
-          <Alert>Nic nie robię. Jestem tutaj dla wyglądu 😊</Alert>
-        </Snackbar>
+        <BottomNav />
       </Grid>
     </Router>
   );
