--- frontend/src/styles/main.scss
@@ -1,14 +1,0 @@
-.center {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-  min-height: 100vh;
-  background-color: #eee;
-  box-sizing: border-box;
-}
-
-body {
-  margin: 0;
-}
